+++ tests/testthat/test-plot_fun.R
@@ -12,14 +12,14 @@ hist_plot <- plot(res)
 hist_plot_mean <- plot(res,vline="mean")
 box_plot <- plot(res,type="box")
 
+output_console <- capture.output(print(res))
+
 test_that("Plot returns a ggplot2 object", {
   expect_equal(class(hist_plot)[2], "ggplot")
   expect_equal(class(hist_plot_mean)[2], "ggplot")
   expect_equal(class(box_plot)[2], "ggplot")
 })
 
-
-test_that("Plot throws error for wrong imputs", {
-  expect_error(plot(res,type="centoids2"), "type must bei either 'hist' or 'box'")
-  expect_error(plot(res,vline="mean2"), "vline must be either 'centroids' or 'mean'")
+test_that("Print returns an output to the console", {
+  expect_equal(output_console[5],"| -0.0277586| -0.0267396|       1|")
 })