+++ NEWS.md
@@ -1,3 +1,5 @@
+# ClustImpute (development version)
+
 # ClustImpute 0.1.7
 
 * Removed dependency from psych package.